WK1/1 - Plus Minus (go)  (#102)
A good way to put the learning of a new language into practice.

Redoing all of the challenges that I've done so far, starting from scratch, in Go.

This one was a larger than normal PR as it had to setup the actions, update the readme and decide on a format for the test files to go into plain text files.