W1/2 - Mini Max Sum (go) (#105)
What a stonker. It is apparent to me just how far I have to go in order
to get very comfortable with Go.

### Workspaces

Lots of time spent figuring out Go workspaces. But I got there.

### LLM

A lot of reliance on asking ChatCPT about X, Y and Z. To be fair, most
(not all, but most) was related to standard package stuff as the online
docs are hard to find stuff about. That's my fault, not the docs.

### HackerRank

I don't know if they do this deliberately, but it feels like it is
intentionally badly put together. Their test running formatted
everything as `int32` but their spec says `1 < arr[i] < **10(9)**`....
int32 doesn't fit 10 to the power of nine.

Either this was a sneaky trick and if it was, may god help me when it
gets advanced, or:

a) it really is very janky
b) I am hot garbage
c) it's a bit janky AND but I've got a way to go with a positive
outlook.



https://www.hackerrank.com/challenges/three-month-preparation-kit-mini-max-sum/problem